Added the script
The scripts monitors a defined deploy using the Octopus Deployment API
and warns when all deploys are finished.

Script thought meanwhile I was looking to the screen waiting for a
deploy to finish.
It could have been done in many languages, but I like doing some
shellscript and I want to train it.